requirements.txt
fastapi
uvicorn
starlette
aiohttp 
sqlalchemy 
databases 
aiosqlite
pytest-asyncio
requests